requirements.txt
discord.py==2.4.0
samp_client==3.0.1
